# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "My API",
        "version": "1.0"
    },
    "paths": {
...
        "/api/inquiries": {
            "get": {
                "tags": [
                    "Inquiry"
                ],
                "summary": "Get Inquiry details",
                "description": "Retrieve the information of the inquiry.",
                "operationId": "inquiry",
                "responses": {
                    "200": {
                        "description": "Inquiry details retrieved successfully.",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated."
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Inquiries"
                ],
                "summary": "Create a new inquiry",
                "operationId": "3f728b15f306ab8aecbf3a468e2a4a56",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Inquiry created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Inquiry created successfully"
                                        },
                                        "inquiry": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Validation failed"
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/inquiry-approved-offers": {
            "get": {
                "tags": [
                    "Inquiry"
                ],
                "summary": "Get Inquiry details",
                "description": "Retrieve the information of the inquiry.",
                "operationId": "approved_inquiry",
                "responses": {
                    "200": {
                        "description": "Inquiry details retrieved successfully.",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated."
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/inquiry-cancellation-offers": {
            "get": {
                "tags": [
                    "Inquiry"
                ],
                "summary": "Get Inquiry details",
                "description": "Retrieve the information of the inquiry.",
                "operationId": "cancellation_inquiry",
                "responses": {
                    "200": {
                        "description": "Inquiry details retrieved successfully.",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated."
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/inquiries/{id}": {
            "get": {
                "tags": [
                    "Inquiries"
                ],
                "summary": "Get details of a specific Inquiry",
                "operationId": "8f67c820771f0052915eb14aa7700dcc",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the inquiry to retrieve",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Inquiry details retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Inquiry retrieved successfully."
                                        },
                                        "inquiry": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Inquiry not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Inquiry not found."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Inquiries"
                ],
                "summary": "Update an existing inquiry",
                "operationId": "8f424ce6fa53d5b9061c81fac0dbbac1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the inquiry to update",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Inquiry updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Inquiry updated successfully"
                                        },
                                        "inquiry": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Inquiry not found"
                    }
                }
            }
        },
...
        "/api/inquries/{id}": {
            "delete": {
                "tags": [
                    "Inquiries"
                ],
                "summary": "Delete a inquiry",
                "operationId": "b5fb0963fa6a9e2f0757a030a3328958",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the inquiry to delete",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Inquiry deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Inquiry deleted successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Inquiry not found"
                    }
                }
            }
        },
...
    },
...
}